# dms-rest/src/main/resources/openapi.yml
openapi: "3.0.0"
info:
  title: Document Management System API
  version: 0.0.1

servers:
  - url: http://localhost:8080

paths:
  /api/v1/status:
    get:
      tags: [ Status ]
      summary: API Health Status
      operationId: getStatus
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusDto'
  /api/v1/documents:
    get:
      tags: [ Document ]
      summary: Get all documents
      operationId: getDocuments
      responses:
        "200":
          description: Documents retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
    post:
      tags: [ Document ]
      summary: Upload a document
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MultipartFile'
      responses:
        "201":
          description: Document created
          headers:
            Location:
              description: URI of the document
              schema:
                type: string
                format: uri
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /api/v1/documents/{id}:
    get:
      tags: [ Document ]
      summary: Get document by id
      operationId: getDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Document retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        "404": { description: Not Found }
    delete:
      tags: [ Document ]
      summary: Delete document by id
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204": { description: Document deleted }
        "404": { description: Not Found }
  /api/v1/documents/{id}/download:
    get:
      tags: [ Document ]
      summary: Download document file by id
      operationId: downloadDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: inline
          in: query
          required: false
          schema:
            type: boolean
          description: If true, serve the file inline for browser preview instead of attachment download
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404": { description: Not Found }
  /api/v1/documents/search:
    get:
      tags: [ Document ]
      summary: Search documents using Elasticsearch
      operationId: searchDocuments
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: scope
          in: query
          required: false
          description: Search scope, e.g. content, name, or all
          schema:
            type: string
            enum: [ content, name, all ]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'

components:
  schemas:
    StatusDto:
      type: object
      properties:
        status:
          type: string
          example: "Alive."

    MultipartFile:
      type: object
      required: [ file ]
      properties:
        file:
          type: string
          format: binary

    DocumentResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        summary:
          type: string
        file:
          $ref: "#/components/schemas/DocumentFileResponse"
        status:
          type: array
          items:
            $ref: "#/components/schemas/DocumentStatusHistoryResponse"

    DocumentFileResponse:
      type: object
      properties:
        objectKey:
          type: string
        name:
          type: string
        type:
          type: string
        size:
          type: number
        pageCount:
          type: integer
        creationDate:
          type: string
          format: date-time
        modificationDate:
          type: string
          format: date-time

    DocumentStatusHistoryResponse:
      type: object
      properties:
        status:
          type: string
        changeDate:
          type: string
          format: date-time

    DocumentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        objectKey:
          type: string
        fileName:
          type: string
        fileSize:
          type: number
          format: double
        fileSizeUnit:
          type: string
        fileType:
          type: string
        fileExtension:
          type: string
        docPageCount:
          type: integer
          format: int64
        docTitle:
          type: string
        docAuthor:
          type: string
        uploaded:
          type: boolean
        scanned:
          type: boolean
        summarized:
          type: boolean
        indexed:
          type: boolean
        docCreatedAt:
          type: string
          format: date-time
        docUpdatedAt:
          type: string
          format: date-time
        summary:
          type: string
          nullable: true
        insertedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string